% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_materials.R
\name{use_activity}
\alias{use_activity}
\title{Copy an activity folder to a selected folder}
\usage{
use_activity(id, destdir = NULL, overwrite = FALSE, open = TRUE)
}
\arguments{
\item{id}{Activity ID, such as \code{"1.2"}, \code{"3.4"}, \code{"final"}, or \code{"final class"}.}

\item{destdir}{Destination directory. If \code{NULL}, students will be asked to choose a folder in RStudio.}

\item{overwrite}{If \code{TRUE}, overwrite an existing folder with the same name.}

\item{open}{If \code{TRUE}, try to open the copied \code{.qmd} file in RStudio.}
}
\value{
Invisibly returns the path to the copied folder.
}
\description{
Copy an activity folder to a selected folder
}
