% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_materials.R
\name{use_lab}
\alias{use_lab}
\title{Copy a lab template to a selected folder}
\usage{
use_lab(id, destdir = NULL, overwrite = FALSE, open = TRUE)
}
\arguments{
\item{id}{Lab ID, such as \code{"1.2"} or \code{"2.1"}.}

\item{destdir}{Destination directory. If \code{NULL}, students will be asked to choose a folder in RStudio.}

\item{overwrite}{If \code{TRUE}, overwrite an existing file with the same name.}

\item{open}{If \code{TRUE}, try to open the copied file in RStudio.}
}
\value{
Invisibly returns the path to the copied file.
}
\description{
Copy a lab template to a selected folder
}
