% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_materials.R
\name{use_lecture}
\alias{use_lecture}
\title{Copy a lecture folder to a selected folder}
\usage{
use_lecture(id, destdir = NULL, overwrite = FALSE, open = TRUE)
}
\arguments{
\item{id}{Lecture ID, such as \code{"3.4"} or \code{"4.1"}.}

\item{destdir}{Destination directory. If \code{NULL}, students will be asked to choose a folder in RStudio.}

\item{overwrite}{If \code{TRUE}, overwrite an existing folder with the same name.}

\item{open}{If \code{TRUE}, try to open the copied \code{.qmd} file in RStudio.}
}
\value{
Invisibly returns the path to the copied folder.
}
\description{
Copy a lecture folder to a selected folder
}
